refactor: rating into a component
Refactors the rating into a reusable component. This cleans up the
commission details view and will make it easy to display the rating
on other pages, such as the profile.
