+++ src/app/pages/commission-board/commission-details/commission-details.component.scss
@@ -25,16 +25,6 @@
     width: auto !important;
 }
 
-.buyer-name {
-    display: flex;
-    align-items: center;
-}
-
-.rating {
-    display: flex;
-    align-items: center;
-}
-
 .card-actions {
     display: flex;
 }
@@ -73,30 +63,16 @@
     align-items: center;
 }
 
-.candidate-name {
-    display: flex;
-    align-items: center;
-}
-
 .crafter-container {
     flex: 1 1 auto;
     .crafter {
         display: flex;
         align-items: center;
         justify-content: center;
-        .crafter-avatar {
-            width: 40px;
-            height: 40px;
-            border-radius: 40px;
-            margin-left: 5px;
-        }
-        .crafter-name {
-            display: flex;
-            flex-direction: column;
-            .name-and-badge {
-                display: flex;
-                align-items: center;
-            }
-        }
+        height: 100%;
     }
+}
+
+.candidates {
+    margin-bottom: 20px;
 }
