+++ src/app/pages/commission-board/rating/rating.component.html
@@ -1,0 +1,21 @@
+<div class="user">
+    <img src="{{character.avatar}}" alt="" class="avatar">
+    <span class="name-and-badge">
+        {{character.name}}
+        <mat-icon *ngIf="character.user.verified">verified_user</mat-icon>
+    </span>
+    <span class="rating">
+        <span *ngFor="let star of getStars(character.user)">
+            <mat-icon *ngIf="star === 1">
+                star
+            </mat-icon>
+            <mat-icon *ngIf="star === 0.5">
+                star_half
+            </mat-icon>
+            <mat-icon *ngIf="star === 0">
+                star_border
+            </mat-icon>
+        </span>
+        <span>({{character.user.ratingsLength}})</span>
+    </span>
+</div>
