+++ src/app/pages/commission-board/rating/rating.component.scss
@@ -1,0 +1,15 @@
+.user {
+    display: flex;
+    flex-direction: row;
+    .avatar {
+        width: 40px;
+        height: 40px;
+        border-radius: 40px;
+        margin-left: 5px;
+    }
+    .name-and-badge {
+        display: flex;
+        align-items: center;
+        margin: 0px 10px;
+    }
+}
